Inital version of !time command
Syntax: !time [area]

Uses the google geocoding api to turn an area specification into a
lat/long coordinate pair, which is then used to get a timezone record to
determine the local time.

Usecase: when working in international groups it's convenient to know if
your co-worker is likely to be awake or asleep.