Make matches case insensitive
This change makes command matching case insensitive. Case sensitivity is a problem for mobile Telegram clients that most likely use auto-correct and capitalisation, and often causes users to have to repeat messages to correct them. As far as I know from my testing, this shouldn't cause any visible changes to the string as it is only lowered for the purpose of matching, and commands such as !addquote won't be lowered in the final output.